Fix migration bugs

+++ database/migrations/2020_09_07_201932_create_posts_table.php
@@ -17,7 +17,7 @@ class CreatePostsTable extends Migration
         
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->string('post_title')->nullable();
             $table->text('post_content')->nullable();
             $table->timestamps();
