+++ database/migrations/2020_09_08_214455_create_categories_table.php
@@ -14,10 +14,10 @@ class CreateCategoriesTable extends Migration
     public function up()
     {
         Schema::create('categories', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('slug')->unique();
-            $table->integer('parent_id')->unsigned()->index()->nullable();
+            $table->bigInteger('parent_id')->unsigned()->index()->nullable();
             $table->timestamps();
 
             $table->foreign('parent_id')->references('id')->on('categories');
